+++ Packages/Tracking/Core/Runtime/Scripts/MRTK Support/LeapMRTKSubsystem.cs
@@ -231,12 +231,12 @@ namespace Leap.Unity.MRTK
                     return;
                 }
 
-                LeapProvider leapProvider = GameObject.FindAnyObjectByType<LeapProvider>();
+                LeapServiceProvider leapProvider = GameObject.FindAnyObjectByType< LeapServiceProvider>();
 
                 // If there is no leap provider in the scene
                 if (leapProvider == null)
                 {
-                    Debug.Log("There are no LeapProviders in the scene, automatically assigning a LeapXRServiceProvider for use with Ultraleap Subsystem for MRTK");
+                    Debug.Log("There are no LeapServiceProviders in the scene, automatically assigning a LeapXRServiceProvider for use with Ultraleap Subsystem for MRTK");
 
 
                     GameObject leapProviderGO = new GameObject("LeapXRServiceProvider");
@@ -247,7 +247,7 @@ namespace Leap.Unity.MRTK
                 }
                 else
                 {
-                    Debug.Log("Ultraleap Subsystem for MRTK is using the existing LeapProvider found in the current scene");
+                    Debug.Log("Ultraleap Subsystem for MRTK is using the existing  LeapServiceProvider found in the current scene");
                 }
 
                 foreach (var hand in hands)